+++ packages/go_router/test/matching_test.dart
@@ -35,37 +35,35 @@ void main() {
           const Placeholder(),
     );
     final Map<String, String> params1 = <String, String>{};
-    final RouteMatch match1 = RouteMatch.match(
+    final List<RouteMatchBase> match1 = RouteMatchBase.match(
       route: route,
-      remainingLocation: '/page-0',
-      matchedLocation: '',
-      matchedPath: '',
+      uri: Uri.parse('/page-0'),
+      rootNavigatorKey: GlobalKey<NavigatorState>(),
       pathParameters: params1,
-    )!;
+    );
 
     final Map<String, String> params2 = <String, String>{};
-    final RouteMatch match2 = RouteMatch.match(
+    final List<RouteMatchBase> match2 = RouteMatchBase.match(
       route: route,
-      remainingLocation: '/page-0',
-      matchedLocation: '',
-      matchedPath: '',
+      uri: Uri.parse('/page-0'),
+      rootNavigatorKey: GlobalKey<NavigatorState>(),
       pathParameters: params2,
-    )!;
+    );
 
     final RouteMatchList matches1 = RouteMatchList(
-      matches: <RouteMatch>[match1],
+      matches: match1,
       uri: Uri.parse(''),
       pathParameters: params1,
     );
 
     final RouteMatchList matches2 = RouteMatchList(
-      matches: <RouteMatch>[match2],
+      matches: match2,
       uri: Uri.parse(''),
       pathParameters: params2,
     );
 
     final RouteMatchList matches3 = RouteMatchList(
-      matches: <RouteMatch>[match2],
+      matches: match2,
       uri: Uri.parse('/page-0'),
       pathParameters: params2,
     );