+++ packages/webview_flutter/webview_flutter_android/lib/src/android_webview_controller.dart
@@ -212,17 +212,13 @@ class AndroidWebViewController extends PlatformWebViewController {
               case ConsoleMessageLevel.debug:
               case ConsoleMessageLevel.tip:
                 logLevel = JavaScriptLogLevel.debug;
-                break;
               case ConsoleMessageLevel.error:
                 logLevel = JavaScriptLogLevel.error;
-                break;
               case ConsoleMessageLevel.warning:
                 logLevel = JavaScriptLogLevel.warning;
-                break;
               case ConsoleMessageLevel.unknown:
               case ConsoleMessageLevel.log:
                 logLevel = JavaScriptLogLevel.log;
-                break;
             }
 
             callback(JavaScriptConsoleMessage(
@@ -754,13 +750,10 @@ class FileSelectorParams {
     switch (params.mode) {
       case android_webview.FileChooserMode.open:
         mode = FileSelectorMode.open;
-        break;
       case android_webview.FileChooserMode.openMultiple:
         mode = FileSelectorMode.openMultiple;
-        break;
       case android_webview.FileChooserMode.save:
         mode = FileSelectorMode.save;
-        break;
     }
 
     return FileSelectorParams(