+++ script/tool/test/podspec_check_command_test.dart
@@ -143,7 +143,6 @@ void main() {
                     .path,
                 '--configuration=Debug',
                 '--skip-tests',
-                '--allow-warnings',
                 '--use-modular-headers',
                 '--use-libraries'
               ],
@@ -159,7 +158,6 @@ void main() {
                     .path,
                 '--configuration=Debug',
                 '--skip-tests',
-                '--allow-warnings',
                 '--use-modular-headers',
               ],
               packagesDir.path),
@@ -201,7 +199,6 @@ void main() {
                     .path,
                 '--configuration=Debug',
                 '--skip-tests',
-                '--allow-warnings',
                 '--use-modular-headers',
                 '--use-libraries'
               ],
@@ -217,7 +214,6 @@ void main() {
                     .path,
                 '--configuration=Debug',
                 '--skip-tests',
-                '--allow-warnings',
                 '--use-modular-headers',
               ],
               packagesDir.path),